--- a/index.d.ts
+++ b/index.d.ts
@@ -1,0 +1,13 @@
+import { LogHandler } from './index'
+
+declare global {
+  namespace NodeJS {
+    interface Global {
+      _logga?: LogHandler[]
+    }
+  }
+
+  interface Window {
+    _logga?: LogHandler[]
+  }
+}