--- a/package.json
+++ b/package.json
@@ -23,6 +23,7 @@
     "build": "npm run build:node && npm run build:browser",
     "build:node": "microbundle build --target node --format es,cjs -o dist/lib",
     "build:browser": "microbundle build --name logga --format es,umd -o dist/browser",
+    "bench": "npm run build:node && node benchmark.js 2>/dev/null",
     "docs": "typedoc --mode file --out docs index.ts && cp *.png docs/"
   },
   "repository": {
@@ -39,13 +40,30 @@
   },
   "homepage": "https://github.com/stencila/logga#readme",
   "devDependencies": {
+<<<<<<< master
     "@stencila/dev-config": "1.4.129",
     "@types/jest": "26.0.20",
+=======
+    "@stencila/dev-config": "1.4.119",
+    "@types/jest": "26.0.19",
+    "benchmark": "2.1.4",
+    "bole": "4.0.0",
+    "bunyan": "1.8.14",
+>>>>>>> perf(Benchmarks): Add benchmarking script
     "jest": "26.6.3",
+    "log4js": "6.3.0",
+    "loglevel": "1.7.1",
     "microbundle": "0.13.0",
+    "pino": "6.10.0",
     "ts-jest": "26.4.4",
+<<<<<<< master
     "typedoc": "0.20.14",
     "typescript": "4.1.3"
+=======
+    "typedoc": "0.19.2",
+    "typescript": "4.1.3",
+    "winston": "3.3.3"
+>>>>>>> perf(Benchmarks): Add benchmarking script
   },
   "jest": {
     "preset": "ts-jest",