I was recently evaluating whether we should continue to maintain Logga, or instead switch to using someone else's logging library. I decided that Logga has a useful combination of features that other libraries do not, and that we should indeed maintain it (but perhaps improve aspects).

As part of that evaluation I implemented some benchmarks and noticed that Logga had poor performance relative to other loggers:

```
bole x 667,896 ops/sec ±0.35% (93 runs sampled)
bunyan x 243,104 ops/sec ±0.27% (89 runs sampled)
logga x 28,811 ops/sec ±0.18% (97 runs sampled)
log4js x 155,397 ops/sec ±0.21% (96 runs sampled)
loglevel x 299,655 ops/sec ±0.38% (95 runs sampled)
pino x 664,888 ops/sec ±0.18% (94 runs sampled)
winston x 232,827 ops/sec ±4.53% (81 runs sampled)
```
It turned out that this was due to automatic stack generation feature (done for logged errors which are not passed an error). After removal of this "feature", performance was at least comparable to others:

```
bole x 657,008 ops/sec ±0.90% (93 runs sampled)
bunyan x 250,412 ops/sec ±0.27% (94 runs sampled)
logga x 253,980 ops/sec ±0.58% (97 runs sampled)
log4js x 143,212 ops/sec ±0.96% (93 runs sampled)
loglevel x 284,880 ops/sec ±0.39% (96 runs sampled)
pino x 671,490 ops/sec ±0.40% (93 runs sampled)
winston x 216,877 ops/sec ±5.55% (78 runs sampled)
```

I continued on the optimization train (see commits for details) until I got performance on par with the fastest of the other libraries:

```
bole x 793,786 ops/sec ±0.42% (92 runs sampled)
bunyan x 270,705 ops/sec ±0.41% (91 runs sampled)
logga x 798,245 ops/sec ±0.30% (94 runs sampled)
log4js x 213,667 ops/sec ±0.30% (92 runs sampled)
loglevel x 598,083 ops/sec ±0.37% (95 runs sampled)
pino x 696,967 ops/sec ±0.51% (92 runs sampled)
winston x 235,126 ops/sec ±6.26% (74 runs sampled)
```

Some time in the future I'd like to do some changes to the API and further optimizations may be possible along side those.