@nokome 2021-01-12
> ...but also had a question regarding the upgrade process for consuming packages/apps. Would they require changes during update?

The API has not changed, only the internals, so no, changes should not be necessary. Regardless, this is a big change so I think it warrants a major version number